Improve handling of enum variants
