+++ macros/src/types/mod.rs
@@ -3,8 +3,10 @@ use quote::quote;
 use syn::{Fields, Generics, ItemEnum, ItemStruct, Result, Variant};
 
 use crate::attr::{EnumAttr, FieldAttr, Inflection, StructAttr};
+use crate::types::generics::format_type;
 use crate::DerivedTS;
 
+mod generics;
 mod named;
 mod newtype;
 mod tuple;
@@ -100,18 +102,28 @@ fn format_variant(
         _ => {}
     };
 
-    let derived_type = type_def(&name, &None, &variant.fields, &Generics::default())?;
-    let inline_type = derived_type.inline;
-    let derived_dependencies = derived_type.dependencies;
+    let generics = Generics::default();
+    let variant_type = type_def(&name, &None, &variant.fields, &generics)?;
+    let variant_dependencies = variant_type.dependencies;
+    let inline_type = variant_type.inline;
 
     formatted_variants.push(match &enum_attr.untag {
         true => quote!(#inline_type),
         false => match &enum_attr.tag {
             Some(tag) => match &enum_attr.content {
-                Some(content) => {
-                    quote!(format!("{{{}: \"{}\", {}: {}}}", #tag, #name, #content, #inline_type))
-                }
-                None => match derived_type.inline_flattened {
+                Some(content) => match &variant.fields {
+                    Fields::Unnamed(unnamed) if unnamed.unnamed.len() == 1 => {
+                        let formatted_type = format_type(&unnamed.unnamed[0].ty, dependencies, &generics);
+                        quote!(
+                            format!("{{{}: \"{}\", {}: {}}}", #tag, #name, #content, #formatted_type)
+                        )
+                    }
+                    Fields::Unit => quote!(format!("{{{}: \"{}\"}}", #tag, #name)),
+                    _ => quote!(
+                        format!("{{{}: \"{}\", {}: {}}}", #tag, #name, #content, #inline_type)
+                    ),
+                },
+                None => match variant_type.inline_flattened {
                     Some(inline_flattened) => quote! {
                         format!(
                             "{{\n{}{}: \"{}\",\n{}\n}}",
@@ -122,7 +134,7 @@ fn format_variant(
                         )
                     },
                     None => {
-                        dependencies.push(quote!(dependencies.append(&mut #derived_dependencies);));
+                        dependencies.push(quote!(dependencies.append(&mut #variant_dependencies);));
                         quote! {
                             format!(
                                 "\n{}{{ {}: \"{}\" }} & {}",
